_id optional; dot-get to meteor package
Meteor/collection2 seems to demand that _id be an optional field if it's
specified; and the Oli node implementation needs _id to be present.  So
mark it optional.

bower doesn't work in meteor packages
(https://github.com/mquandalle/meteor-bower/issues/3), so convert
dot-get to a meteor package and use it that way.